refactor: Move classes into appropriate packages
