+++ app/src/main/java/com/example/stren/app/StrenAppViewModel.kt
@@ -1,27 +1,32 @@
-package com.example.stren
+package com.example.stren.app
 
+import android.util.Log
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import com.example.stren.core.service.AuthenticationService
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
+private const val TAG = "StrenAppViewModel"
+
 @HiltViewModel
 class StrenAppViewModel @Inject constructor(
     authenticationService: AuthenticationService
-) :
-    ViewModel() {
+) : ViewModel() {
+    var isUserSignedIn = mutableStateOf(false)
+        private set
 
     init {
         authenticationService.addAuthStateListeners(
             onUserAuthenticated = {
                 isUserSignedIn.value = true
+                Log.d(TAG, "stateListen - isUserSignedIn: ${isUserSignedIn.value}")
             },
             onUserNotAuthenticated = {
                 isUserSignedIn.value = false
-            })
-    }
+                Log.d(TAG, "stateListen - isUserSignedIn: ${isUserSignedIn.value}")
 
-    var isUserSignedIn = mutableStateOf(false)
-        private set
+            })
+        Log.d(TAG, "init() - isUserSignedIn: ${isUserSignedIn.value}")
+    }
 }
