+++ app/src/main/java/com/example/stren/app/StrenApplication.kt
@@ -1,4 +1,4 @@
-package com.example.stren
+package com.example.stren.app
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
